requirements.txt
google-generativeai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
gspread
python-dotenv
requests
